Move Window/DialogWindow/ComposePanel layout-direction tests to a separate file.
