+++ compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/awt/ComposeDialog.desktop.kt
@@ -35,7 +35,7 @@ import java.awt.Window
 import java.awt.event.MouseListener
 import java.awt.event.MouseMotionListener
 import java.awt.event.MouseWheelListener
-import java.util.*
+import java.util.Locale
 import javax.swing.JDialog
 import org.jetbrains.skiko.SkiaLayerAnalytics
